work in progress

+++ src/main/java/nl/digitalekabeltelevisie/data/mpeg/psi/TableSection.java
@@ -515,7 +515,7 @@ public class TableSection implements TreeNode{
 		this.nextVersion = next;
 	}
 
-	public long getFirst_packet_no() {
+	public int getFirst_packet_no() {
 		return firstPacketNo;
 	}
